health_monitor/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
motor==3.3.2
python-dotenv==1.0.1
